+++ package.json
@@ -24,8 +24,8 @@
     "@alcalzone/release-script-plugin-lerna": "~3.5.0",
     "@alcalzone/release-script-plugin-license": "~3.5.0",
     "@alcalzone/release-script-plugin-manual-review": "~3.5.8",
-    "@iobroker/types-dev": "file:packages/types-dev",
     "@iobroker/types": "file:packages/types-public",
+    "@iobroker/types-dev": "file:packages/types-dev",
     "@microsoft/api-extractor": "^7.33.2",
     "@tsconfig/node12": "^1.0.11",
     "@types/chai": "^4.3.3",
@@ -67,12 +67,12 @@
     "semver": "^7.3.7",
     "sinon": "^15.0.0",
     "sinon-chai": "^3.7.0",
+    "ts-json-schema-generator": "~1.0.0",
     "ts-node": "^10.9.1",
     "typedoc": "^0.23.11",
     "typedoc-plugin-markdown": "^3.13.5",
     "typedoc-plugin-missing-exports": "^1.0.0",
-    "typescript": "~4.8.4",
-    "typescript-json-schema": "^0.55.0"
+    "typescript": "~4.8.4"
   },
   "scripts": {
     "clean": "rimraf packages/*/build packages/*/*.tsbuildinfo",
@@ -81,9 +81,8 @@
     "build:types": "npm run build --workspace=@iobroker/types",
     "watch:ts": "npm run build:ts -- --watch",
     "build": "npm run build:ts && npm run build:types",
-    "postbuild": "lerna run postbuild",
+    "postbuild": "lerna run postbuild && npm run update-schema",
     "preinstall": "lerna run preinstall",
-    "postbuild": "npm run update-schema",
     "install": "lerna run install",
     "release": "release-script",
     "update-schema": "ts-node schemas/updateSchemas.ts",
