+++ packages/controller/conf/iobroker-dist.json
@@ -90,7 +90,7 @@
             "files": 24,
             "// files": "Minimal number of backup files, after the deletion will be executed according to backupTime settings",
             "hours": 48,
-            "// hoursC": "All backups older than 48 hours will be deleted. But only if the number of files is greater than of backupNumber",
+            "// hours": "All backups older than 48 hours will be deleted. But only if the number of files is greater than of backupNumber",
             "period": 120,
             "// period": "by default backup every 2 hours. Time is in minutes. To disable backup set the value to 0",
             "path": "",
@@ -163,5 +163,6 @@
 
     },
     "// dnsResolution": "Use 'verbatim' for ipv6 first, else use 'ipv4first'",
-    "dnsResolution": "ipv4first"
+    "dnsResolution": "ipv4first",
+    "dataDir": ""
 }
