+++ packages/types-dev/package.json
@@ -24,6 +24,7 @@
     },
     "files": [
         "index.d.ts",
+        "config.d.ts",
         "objects.d.ts",
         "LICENSE"
     ]
